Params refactoring - move common params in a base class for code deduplication purposes
