+++ src/main/java/com/nvidia/cuvs/lucene/GPUSearchParams.java
@@ -10,63 +10,12 @@ import com.nvidia.cuvs.CagraIndexParams.CuvsDistanceType;
 import com.nvidia.cuvs.CuVSIvfPqParams;
 import com.nvidia.cuvs.lucene.CuVS2510GPUVectorsWriter.IndexType;
 import java.util.Objects;
-import java.util.function.Supplier;
 
-public class GPUSearchParams {
+public class GPUSearchParams extends ParamsBase {
 
-  public static enum Strategy {
-    /*
-     * This strategy allows for automatic selection of the underlining CAGRA build algorithm.
-     * With this strategy we use NN_DESCENT for data set less then 5M vectors else we use IVF_PQ.
-     * Indexing parameters, especially for IVF_PQ, are heuristically identified automatically.
-     *
-     * This is the default and the recommended strategy.
-     */
-    HEURISTIC,
-    /*
-     * This is an option when the end-user would want to use custom parameter values.
-     * This strategy should only be used under expert guidance.
-     */
-    CUSTOM
-  }
-
-  /*
-   * TODO: Update boundaries for all parameters when a consensus is reached.
-   * Issue: https://github.com/rapidsai/cuvs-lucene/issues/99
-   */
-  public static final int MIN_WRITER_THREADS = 1;
-  public static final int MAX_WRITER_THREADS = 512;
-  public static final int MIN_INT_GRAPH_DEG = 2;
-  public static final int MAX_INT_GRAPH_DEG = 512;
-  public static final int MIN_GRAPH_DEG = 1;
-  public static final int MAX_GRAPH_DEG = 512;
-  public static final int MIN_NN_DESCENT_NUM_ITERATIONS = 1;
-  public static final int MAX_NN_DESCENT_NUM_ITERATIONS = 100;
-
-  public static final int DEFAULT_INT_GRAPH_DEGREE = 128;
-  public static final int DEFAULT_GRAPH_DEGREE = 64;
-  public static final CagraGraphBuildAlgo DEFAULT_CAGRA_GRAPH_BUILD_ALGO =
-      CagraGraphBuildAlgo.NN_DESCENT;
   public static final IndexType DEFAULT_INDEX_TYPE = IndexType.CAGRA;
-  public static final int DEFAULT_WRITER_THREADS = 32;
-  public static final Strategy DEFAULT_STRATEGY = Strategy.HEURISTIC;
-  public static final CuvsDistanceType DEFAULT_CUVS_DISTANCE_TYPE = CuvsDistanceType.L2Expanded;
-  public static final int DEFAULT_NN_DESCENT_NUM_ITERATIONS = 20;
 
-  public static final Supplier<CuVSIvfPqParams> DEFAULT_IVF_PQ_PARAMS =
-      () -> {
-        return new CuVSIvfPqParams.Builder().build();
-      };
-
-  private final int writerThreads;
-  private final int intermediateGraphDegree;
-  private final int graphdegree;
-  private final CagraGraphBuildAlgo cagraGraphBuildAlgo;
   private final IndexType indexType;
-  private final CuVSIvfPqParams cuVSIvfPqParams;
-  private final Strategy strategy;
-  private final CuvsDistanceType cuvsDistanceType;
-  private final int nnDescentNumIterations;
 
   /**
    * Constructs an instance of {@link GPUSearchParams} with specific parameter values.
@@ -91,53 +40,21 @@ public class GPUSearchParams {
       CuVSIvfPqParams cuVSIvfPqParams,
       Strategy strategy,
       CuvsDistanceType cuvsDistanceType,
-      int nnDescentNumIterations) {
-    super();
-    this.writerThreads = writerThreads;
-    this.intermediateGraphDegree = intermediateGraphDegree;
-    this.graphdegree = graphdegree;
-    this.cagraGraphBuildAlgo = cagraGraphBuildAlgo;
+      int nnDescentNumIterations,
+      int maxConn,
+      int beamWidth) {
+    super(
+        writerThreads,
+        intermediateGraphDegree,
+        graphdegree,
+        maxConn,
+        beamWidth,
+        cagraGraphBuildAlgo,
+        cuVSIvfPqParams,
+        strategy,
+        cuvsDistanceType,
+        nnDescentNumIterations);
     this.indexType = indexType;
-    this.cuVSIvfPqParams = cuVSIvfPqParams;
-    this.strategy = strategy;
-    this.cuvsDistanceType = cuvsDistanceType;
-    this.nnDescentNumIterations = nnDescentNumIterations;
-  }
-
-  /**
-   * Get the cuVS writer threads parameter
-   *
-   * @return cuVS writer threads parameter
-   */
-  public int getWriterThreads() {
-    return writerThreads;
-  }
-
-  /**
-   * Get the intermediate graph degree
-   *
-   * @return the graph degree parameter
-   */
-  public int getIntermediateGraphDegree() {
-    return intermediateGraphDegree;
-  }
-
-  /**
-   * Get the graph degree
-   *
-   * @return the graph degree parameter
-   */
-  public int getGraphdegree() {
-    return graphdegree;
-  }
-
-  /**
-   * Get the CAGRA build algorithm parameter value
-   *
-   * @return the CAGRA build algorithm parameter value
-   */
-  public CagraGraphBuildAlgo getCagraGraphBuildAlgo() {
-    return cagraGraphBuildAlgo;
   }
 
   /**
@@ -147,69 +64,6 @@ public class GPUSearchParams {
    */
   public IndexType getIndexType() {
     return indexType;
-  }
-
-  /**
-   * Get the instance of CuVSIvfPqParams
-   *
-   * @return an instance of CuVSIvfPqParams
-   */
-  public CuVSIvfPqParams getCuVSIvfPqParams() {
-    return cuVSIvfPqParams;
-  }
-
-  /**
-   * Get the chosen strategy:
-   *
-   * When HEURISTIC [Default] is chosen, the CAGRA build algorithm and its indexing parameters are automatically chosen based on the size of the data set
-   * When CUSTOM is chosen, the build algorithm and its parameters (either defaults or overridden values with the use of With* methods) is used internally
-   *
-   *
-   * @return get the chosen {@link Strategy}
-   */
-  public Strategy getStrategy() {
-    return strategy;
-  }
-
-  /**
-   * Get the cuvs distance type
-   *
-   * @return the distance type
-   */
-  public CuvsDistanceType getCuvsDistanceType() {
-    return cuvsDistanceType;
-  }
-
-  /**
-   * get the number of Iterations to run if building with NN_DESCENT
-   *
-   * @return the number of iterations for NN_DESCENT
-   */
-  public int getnNDescentNumIterations() {
-    return nnDescentNumIterations;
-  }
-
-  @Override
-  public String toString() {
-    return "GPUSearchParams [writerThreads="
-        + writerThreads
-        + ", intermediateGraphDegree="
-        + intermediateGraphDegree
-        + ", graphdegree="
-        + graphdegree
-        + ", cagraGraphBuildAlgo="
-        + cagraGraphBuildAlgo
-        + ", indexType="
-        + indexType
-        + ", cuVSIvfPqParams="
-        + cuVSIvfPqParams
-        + ", strategy="
-        + strategy
-        + ", cuvsDistanceType="
-        + cuvsDistanceType
-        + ", nnDescentNumIterations="
-        + nnDescentNumIterations
-        + "]";
   }
 
   /**
@@ -226,6 +80,8 @@ public class GPUSearchParams {
     private Strategy strategy = DEFAULT_STRATEGY;
     private CuvsDistanceType cuvsDistanceType = DEFAULT_CUVS_DISTANCE_TYPE;
     private int nnDescentNumIterations = DEFAULT_NN_DESCENT_NUM_ITERATIONS;
+    private int maxConn = DEFAULT_MAX_CONN;
+    private int beamWidth = DEFAULT_BEAM_WIDTH;
 
     /**
      * Set the number of cuVS writer threads while building the index
@@ -344,56 +200,39 @@ public class GPUSearchParams {
     }
 
     /**
+     * Set the max connections parameter while building HNSW index with fallback mechanism
+     * Valid range - Minimum: {@value MIN_MAX_CONN}, Maximum: {@value MAX_MAX_CONN}
+     * Default value - {@value DEFAULT_MAX_CONN}
+     *
+     * @param maxConn the max connections parameter
+     * @return instance of {@link Builder}
+     */
+    public Builder withMaxConn(int maxConn) {
+      this.maxConn = maxConn;
+      return this;
+    }
+
+    /**
+     * Set the beam width parameter while building HNSW index with fallback mechanism
+     * Valid range - Minimum: {@value MIN_BEAM_WIDTH}, Maximum: {@value MAX_BEAM_WIDTH}
+     * Default value - {@value DEFAULT_BEAM_WIDTH}
+     *
+     * @param beamWidth the beam width parameter
+     * @return instance of {@link Builder}
+     */
+    public Builder withBeamWidth(int beamWidth) {
+      this.beamWidth = beamWidth;
+      return this;
+    }
+
+    /**
      * Validates the input parameters.
      *
      * @throws IllegalArgumentException
      */
     private void validate() throws IllegalArgumentException {
-      if (writerThreads < MIN_WRITER_THREADS || writerThreads > MAX_WRITER_THREADS) {
-        throw new IllegalArgumentException(
-            "writerThreads not in valid range. Valid range: ["
-                + MIN_WRITER_THREADS
-                + ", "
-                + MAX_WRITER_THREADS
-                + "]");
-      }
-      if (intermediateGraphDegree < MIN_INT_GRAPH_DEG
-          || intermediateGraphDegree > MAX_INT_GRAPH_DEG) {
-        throw new IllegalArgumentException(
-            "intermediateGraphDegree not in valid range. Valid range: ["
-                + MIN_INT_GRAPH_DEG
-                + ", "
-                + MAX_INT_GRAPH_DEG
-                + "]");
-      }
-      if (graphdegree < MIN_GRAPH_DEG || graphdegree > MAX_GRAPH_DEG) {
-        throw new IllegalArgumentException(
-            "graphdegree not in valid range. Valid range: ["
-                + MIN_GRAPH_DEG
-                + ", "
-                + MAX_GRAPH_DEG
-                + "]");
-      }
-      if (Objects.isNull(cagraGraphBuildAlgo)) {
-        throw new IllegalArgumentException("cagraGraphBuildAlgo cannot be null.");
-      }
       if (Objects.isNull(indexType)) {
         throw new IllegalArgumentException("indexType cannot be null.");
-      }
-      if (Objects.isNull(strategy)) {
-        throw new IllegalArgumentException("strategy cannot be null.");
-      }
-      if (Objects.isNull(cuvsDistanceType)) {
-        throw new IllegalArgumentException("cuvsDistanceType cannot be null.");
-      }
-      if (nnDescentNumIterations < MIN_NN_DESCENT_NUM_ITERATIONS
-          || nnDescentNumIterations > MAX_NN_DESCENT_NUM_ITERATIONS) {
-        throw new IllegalArgumentException(
-            "nnDescentNumIterations not in valid range. Valid range: ["
-                + MIN_NN_DESCENT_NUM_ITERATIONS
-                + ", "
-                + MAX_NN_DESCENT_NUM_ITERATIONS
-                + "]");
       }
     }
 
@@ -403,9 +242,6 @@ public class GPUSearchParams {
      * @return instance of {@link GPUSearchParams}
      */
     public GPUSearchParams build() {
-      if (Objects.isNull(cuVSIvfPqParams)) {
-        cuVSIvfPqParams = DEFAULT_IVF_PQ_PARAMS.get();
-      }
       validate();
       return new GPUSearchParams(
           writerThreads,
@@ -416,7 +252,9 @@ public class GPUSearchParams {
           cuVSIvfPqParams,
           strategy,
           cuvsDistanceType,
-          nnDescentNumIterations);
+          nnDescentNumIterations,
+          maxConn,
+          beamWidth);
     }
   }
 }
