+++ src/test/java/com/nvidia/cuvs/lucene/TestGPUSearchParams.java
@@ -49,7 +49,7 @@ public class TestGPUSearchParams extends LuceneTestCase {
     assertEquals(DEFAULT_INDEX_TYPE, params.getIndexType());
     assertEquals(DEFAULT_STRATEGY, params.getStrategy());
     assertEquals(DEFAULT_CUVS_DISTANCE_TYPE, params.getCuvsDistanceType());
-    assertEquals(DEFAULT_NN_DESCENT_NUM_ITERATIONS, params.getnNDescentNumIterations());
+    assertEquals(DEFAULT_NN_DESCENT_NUM_ITERATIONS, params.getNNDescentNumIterations());
   }
 
   @Test
